I made a lot of changes
I forgot to commit as I went, I just got my head down and got so hyper focused that I just didn't commit. So here we are. I've done most of the alignment and style, most of the style of the page is now in, except the footer and some cleaning up